pyproject.toml
[tool.poetry]
name = "real-estate-listing-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "real-estate-listing-api"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.0"
pydantic-settings = "^2.1.0"
pymongo = "^4.6.1"
black = "^23.12.0"
mypy = "^1.7.1"
flake8 = "5.0.4"
pytest = "^7.4.3"
pyjwt = "^2.8.0"
pydantic = "^2.5.2"
flask-pydantic = "^0.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
